Add uint8 type, adjust Type module
Type.t has been updated to allow for integers of varying bitwidths. Furthermore,
semantic analysis now assigns integer literals the least powerful type that can
fit the value of the literal, and coerces it upwards as necessary. At binding
sites, such as `let` and `var`, if there's no type annotation, the bound
expression is checked for "promotability"; that is, whether it only contains
integer literals, array literals, and binary operators. If the expression is
promotable, it's type is promoted such that int64 is chosen as the default
integer type.